🎨: 할인 금액 렌더링 함수 추출
